Fix incorrect ack output parsing
As an example, if the line output was

25:  hello-100-some more text

and I was searching for "hello", then everything through the first -
would be considered the file name, and 100 would be considered the line
number. No match for "hello" would be found.

This commit fixes that problem so that 25 gets detected as the line
number, there is no file name detected, and it detects the "hello"
match.

Unfotunately, the fix involves relying on the fact that the line is
color coded, which only started happening with ack 1.94. I haven't been
able to think of a fix that would also work for non-color coded line
numbers, and thus support older versions of ack. Parsing older versions
of ack will work just as it did before, and still have the bug this
commit is meant to fix.