Stop extra window scrolling on previous/next match
By setting compilation-context-lines in ack-find-match we were causing
excessive window scrolling when the compilation-goto-locus function was
called. Grep-mode didn't do this and I find it harder to keep my place
as I'm jumping to different matches. I'm not sure if this was
intentional or not, but it seems better to turn it off.